Don't emit documentClosed repeatedly for the same document
Numerous slots are connected to this signal. They are all invoked
repeatedly and redundantly. The number of the invocations equals one
plus the number of times the closed document was activated via reopening
(e.g. by Most-Recently-Used Document Switcher or by Document View).

Connect the slot DocumentController::notifyDocumentClosed() only when
the document is opened for the first time to ensure that the
documentClosed signal is emitted once. Rename the local boolean variable
that determines whether the document was already open for clarity.