Adapt plugin target/outputnames to match the plugin id (needed with KF6)
KPluginMetaData in KF6 derives the plugin id no longer from the "Id" entry
in the plugin metadata, but instead from the QFileInfo::completeBaseName()
of the plugin binary.
CMake uses the OUTPUT_NAME property of the logical plugin target, which
defaults to the name of the target. This patch thus changes the logical
target names of the plugin module type libraries to match the plugin id,
to achieve consistent plugin ids between the KF5 & KF6 versions.
Plugin ids are e.g. used with the user's configuration of enabled/disabled
plugins.