sublime: optimize updating document list menu
Whenever a view is added to a container (e.g. when a document is opened)
or a document is renamed, activating actions for all open documents in
the same container (i.e. the document's split view) are destroyed and
created again. This means creating and destroying a QAction O(N^2)
times, where N is the number of open documents in the container.

Optimize this by creating and inserting a single action for the
just-added document or moving the renamed document's action.

At the parent commit starting KDevelop with 242 open documents in
syntax-highlighting/autotests/folding/ takes 37 seconds. At this commit
the same operation takes 25 seconds. I hoped this optimization would
noticeably speed up opening the more usual C++ and C files, but this is
not the case. The QAction construction for the syntax-highlighting test
files is particularly slow, because MIME type lookup and especially
QIcon operations happen to be very slow for these documents.

This commit also introduces two minor behavior changes:
* Document titles are now compared case-insensitively to decide whether
  to disambiguate them. Whether this improves user experience is hard to
  tell, but since actions are sorted by title case-insensitively, case-
  insensitive comparison is necessary for correctness. Otherwise, two
  open documents with equal titles, separated by another document with a
  different-case title, would not be disambiguated.
* An action's text is set to the last two segments of QUrl::path(), not
  of QUrl::toString(), for disambiguation. QUrl::path() should be faster
  and showing an URL's slash separators is not useful.