Adapt to invalid begin iterator of QList/6 vs. QVector/5 on erase at begin
Implementation of QList (replacing QVector via a typedef) in Qt6 when
erasing at the begin of the list instead of memmoving data around just
moves the internal pointer to the begin around.
This breaks the assumption of the code written for QVector, caching the
iterator to the start.

Instead always fetch the begin iterator on demand.