MIDebuggerPlugin: setup actions only if there is a QGuiApplication
test_midbus is a UI-less test. In Qt6 the creation of QAction instances
tries to use a QGuiApplication instance and will crash hard if there is
none.

No access to Core::setupFlags in the plugins, only to ICore, so other than
e.g. DebugController::initializeUi() MIDebuggerPlugin cannot check that.
Thus using the same approach as other "are-we-ui" non-kdevelop-core code.