TestCore::shutdown(): deal with Core possibly gone in Qt6's QTest::qWait()
Behaviour of QCoreApplication::processEvents in Qt5 seems to have been
different. There no events ended up being already posted to the event
handlers. Thus the QDeferredDeleteEvent for the Core instance was only
created during QCoreApplication::sendPostedEvents and added to the queue,
to be handled in another 10 ms slot in QTest::qWait.
As only one slot is used, the event then stays in the queue only and thus
the Core instance has been still around.