SourceFileTemplate: fail when cannot open the archive
Now more test functions of test_templateclassgenerator and
test_generationtest fail and fail earlier when opening a KArchive fails
because of the "Could not read tar header" error.

The dialog Create Files from Template no longer shows empty Header and
Implementation previews when opening a KArchive fails. It also no longer
allows to navigate through and fill in file generation pages to
eventually produce an empty file because of this failure.