Un-bump KDevelop minor version
8f61478a020bdb7bba79a0a2879dece34197c531 increased the KDE Gear version
from 23.11 to 24.01. The formula that calculates
KDEVELOP_OWN_MINOR_VERSION based on the release service version did not
anticipate that and bumped the minor version for the second time since
the previous release 5.12.2308 to 14. Adjust the formula to group the
KDE Gear versions x.01 and x.02 with (x-1).11 and (x-1).12 rather than
with x.03 and x.04. The new formula supports two plausible future KDE
Gear versioning schemes:
    * switch to x.01-x.02, x.05-x.06 and x.09-x.10;
    * back to x.03-x.04, x.07-x.08 and x.11-x.12.
If the versions go back to x.03-x.04 etc. but skip the 24.04 release,
the final subtrahend at the end of the expression will have to be
increased to 60 in order not to skip the minor version 14.

Current KDevelop version is now 5.13.240175 instead of 5.14.240175.