grepview: replace Depth comment with tooltip and What's This
c0d4221d344b47f29a7862c4bc086e0a7a8bae7c added a comment to the Depth
spinbox's tooltip instead of tooltip text, probably by mistake. Remove
the useless comment and add a short tooltip and a more detailed
What's This explanation to the spinbox.

The new wordings have the same meaning as the removed comment, but are
also based on the documentation of the -maxdepth option in the `find`
man page and the documentation of the --max-depth option in the
`git-grep` man page. Use the word "directory" rather than "folder",
because it is used in all other tooltips of the Find/Replace in Files
dialog.