GrepFindFilesThread: navigate directories faster in findFiles()
FileFinder::findFiles() computes canonical paths using QFileInfo in
loops and recursively. Introduce a private helper function to compute
the canonical path of each search location once.

Use QDirIterator instead of QDir::entryInfoList() to avoid unnecessary
QList collections.

When a nonexistent single file is the search location, its canonical
file path is empty, but FileFinder::findFiles() does not check that and
puts the empty string into the results list. This causes the following
warning to be printed while trying to open the empty-string result:
    QFSFileEngine::open: No file name specified
Check whether the canonical file path is empty to prevent the warning.