grepview: normalize search location URL path segments
GrepDialog::updateLimitToProjectEnabled() passes path(s) entered by the
user to directoriesInProject(), which forwards the path(s) to
getDirectoryChoice() in order to convert them to a list of URLs.
directoriesInProject() then proceeds to invoke
ProjectController::findProjectForUrl() for each URL in the list.
findProjectForUrl() constructs an IndexedString from the argument URL if
at least one project is open. The QUrl IndexedString constructor asserts
that the URL's path segments are normalized.

Normalize path segments in getDirectoryChoice() to eliminate the
assertion failure when the user enters e.g. an absolute path that ends
with "/." or "/.." as a search location.