test_findreplace: test search limited to project files
Do not add separate test data rows for searches limited and not limited
to project files, because a lot of work, such as temporary directory
setup, can be done once and reused. When the test() lambda fails, it
early-returns, but FindReplaceTest::testIncludeExcludeFilters() does not
care and proceeds to the next iteration of the projectFilesOnly loop.
This way, even if the test not limited to project files fails, the
limited test still runs. One downside of this approach is the reliance
on debug output to pinpoint failures, which can be enabled e.g. via
    QT_LOGGING_RULES='default.debug=true' ./bin/test_findreplace
Another downside is that when both projectFilesOnly loop iterations
fail, QTest reports two failures for a single data row.

Move the createChild() utility from quickopentestbase.h to testproject.h
and reuse it in another added utility addChildrenFromFileSystem().