GrepFindFilesThread: match relative paths against Exclude filter
Two KDevelop users have reported the "No files found matching the
wildcard patterns" bug when the project directory is within a parent
directory named "build". The /build/ pattern is present in the default
Exclude filter. Therefore, all absolute file paths in such a project
directory match the default Exclude filter and are not searched in.
Match paths relative to the current search location in order to fix the
bug.

A downside of this fix is that when multiple projects are open,
excluding one of them from the search is no longer possible. Hopefully
this is a very rare use case.

FileFinder::findFilesCanonical: don't search in a subdirectory whose
relative path matches the Exclude filter. Since the relative path to a
file within a subdirectory contains the relative path to the
subdirectory, it would also match the Exclude filter and therefore would
not be inserted into the results list.

BUG: 361760
FIXED-IN: 5.14