debugger: isolate saved line numbers from the moving cursors
When a document is edited and then closed without saving, a breakpoint
located in the document retains its moving cursor's (tracked) line
number. When this document is then opened again, the breakpoint is
likely misplaced, because its line number depends on discarded text.
This and subsequent commits ensure correct breakpoint line number if
feasible, that is, as long as the document is not modified outside of
KDevelop.

Make all breakpoint lines and URL setters delegate to the general
function Breakpoint::setLocation(). This is in preparation for full
document line tracking support. Crucially, setLocation() now also
updates the moving cursor (if attached).

The tracked document line number is now returned by Breakpoint::line().
Now the saved and tracked line numbers may diverge until the document
is saved. savedLine() always returns the Breakpoint::m_line that is not
updated while the document is being edited. It is instead updated when
the document is saved (saveMovingCursorLine()) or when the breakpoint's
location is assigned directly (setLocation()).

Now it should be harder to break the document line tracking if a
document is closed without saving: the moving cursors in that document
are simply discarded, and the breakpoints revert to their last assigned
m_line numbers. However, there are edge cases where the reversion does
not work well.

Update test_breakpointmodel to allow verifyBreakpoint() checks to
partially succeed. This is now required because b1->line() returns a
correct document line number but checks on b2 cannot yet succeed.