Port IPartController and OpenWithPlugin away from KMimeTypeTrader
Use KParts::PartLoader and KPluginFactory::loadFactory instead.
We depend on new-enough frameworks to do this without any ifdefs.

findPartFactory() and its only caller IPartController::createPart()
now interpret the preferredName argument as a plugin ID
(KPluginMetaData::pluginId) instead of the filename of a service
desktop entry (KService::desktopEntryName). Accordingly,
OpenWithPlugin::openPart() now passes the plugin ID instead of the
desktop entry name to IDocumentController::openDocument().

Igor Kushnir searched and learned that
OpenWithPlugin::openPart() is the only function that calls
DocumentController::openDocument(const QUrl& inputUrl, const QString& prefName),
which in turn is the only caller of DocumentController::openDocumentInternal
with a nonempty prefName, which in turn is the only creator of a PartDocument
with a nonempty preferredPart, which in turn is the only indirect caller of
findPartFactory() with a nonempty preferredName.

Porting the OpenWithPlugin is a bit more invasive than I hoped.
We now need to distinguish explicitly between kparts and external
applications, and thus have to split up the code a bit more.
I've tried to share as much code as possible though, and tested
it all manually to verify it works.