Cleanup PartController::{can,}createPart mimetype/url handling
PartController::createPart is only called from PartDocument after
the DocumentController checks canCreatePart. That part never returned
true for invalid urls - but an empty url is invalid too. As such, we
can simplify the code and assume it's never called with empty URLs.
The logic to map empty urls to text documents is handled specially
in the DocumentController anyhow.