Disable Open With embedded parts for directories
KDevelop can only open files, not folders, in its integrated editor.
We would never actually show the embedded part when the user tries to
do so on a directory. Opening in an external app works just fine.
So for now, we just disable the embedded parts for directories here
to remove this confusing and non-working functionality.