openwith: don't read the same config entry repeatedly
defaultIdForMimeType() is always called with the same argument -
OpenWithPlugin::m_mimeType. The return value of this function depends
only on m_mimeType and the config group "Open With Defaults".

m_mimeType is modified in a single place in
OpenWithPlugin::updateMimeTypeForUrls(). The "Open With Defaults" group
is written to in a single place in
OpenWithPlugin::rememberDefaultChoice().

Replace defaultIdForMimeType() with OpenWithPlugin::m_defaultId to avoid
reading the default ID for the same MIME type repeatedly.