openwith: store file opener ID type in config
Before the recent commit "Port IPartController and OpenWithPlugin away
from KMimeTypeTrader", a service storage ID was always stored as a file
opener ID. The porting commit changed this so that a plugin ID is stored
for a part and a service storage ID for an external application.

OpenWithPlugin::openDefault() attempts to create a service to determine
the stored ID type. This works as long as no plugin ID equals a storage
ID, which is likely but not guaranteed.

Store the file opener ID type along with the ID itself in config by
prefixing a part ID with "PART-ID:". This way, an external application
service storage ID already stored in a user config is recognized
normally. An already stored part service storage ID is now considered
invalid and effectively ignored, because there is no easy way to convert
a service storage ID to a plugin ID.