openwith: validate service storage ID read from config
Since the recent commit "Port IPartController and OpenWithPlugin away
from KMimeTypeTrader", a service storage ID read from config is no
longer validated before use. This has caused several regressions in
functions that use m_defaultOpener when its ID is nonempty but invalid
(that is, when an external application service with such a storage ID
does not exist):
1. canOpenDefault() returns true for the directory MIME type, even if no
   directory-opening service is available in the system.
2. openDefault() no longer falls back to hard-coded default openers.
3. actionsForParts() no longer marks the Default Editor action as
   default.

Some IDs stored in a user config are likely to be nonempty and invalid,
because the aforementioned porting commit also stopped supporting part
service storage IDs.

Look for a service specified by the service storage ID read from config.
If not found, mark the file opener as invalid. This way, an invalid
stored service storage ID is treated equivalently to a not-stored ID.

I could not find an efficient way to validate a part plugin ID. Existing
user config files do not contain plugin IDs, because they are stored
only since the aforementioned porting commit. But when a part ID is
stored and then becomes invalid in the future, KDevelop will misbehave.