openwith: enclose FileOpener in namespace OpenWithUtils
"FileOpener" is a fairly general name that might clash with an unrelated
class in the future. An alternative way to prevent future name clashes
is to make FileOpener a nested class of the class OpenWithPlugin. But
then the two classes would be gratuitously coupled and the rather
general class FileOpener would become more difficult to reuse elsewhere.