debugger: split Breakpoint::setMovingCursor()
setMovingCursor() is called for two distinct purposes: to delete the
old moving cursor if one exists and to assign a newly created moving
cursor. Split it into two specialized functions in order to constrain
the interface and reduce branching when a subsequent commit ties the
breakpoint mark to its moving cursor.