Extract BreakpointModel::setupDocumentBreakpoints()
Currently, the extracted function is called only once. But it will be
reused in another place in a future commit.

Reorder setupDocumentBreakpoints() before connecting to the
markChanged() signal. This does not matter right now. But in a future
commit, setupDocumentBreakpoints() will set up breakpoint marks along
with the moving cursors, which would undesirably invoke
BreakpointModel::markChanged() if connected.