BreakpointModel: preserve document line tracking during reload
The markChanged() signal is repeatedly emitted while a document that
contains breakpoint marks is reloaded. Thus the
BreakpointModel::markChanged() slot is invoked and erroneously removes
and adds breakpoints. This slot is supposed to be inhibited during
programmatic mark changes, but is not inhibited while a document is
reloaded and KTextEditor removes then adds back document marks.
6f12697d36963c8f10a367ebddbcbcd959912779 changed the results of such
unsupported reloading. Without dedicated code that handles reloading,
these results remain messy: existing breakpoints are moved to the first
line of the document, new breakpoints with default enabled, condition
and ignore-hits data appear at the locations of those moved
breakpoints, etc. Reloading a document also invalidates all moving
cursors bound to it, which prevents the document's Breakpoint instances
from tracking their marks going forward.

KTextEditor::Document unconditionally emits the aboutToReload() signal
at the beginning and the reloaded() signal at the end of the reloading
process. Connect to these signals in order to (mostly) fix the
reloading bugs.

A new slot BreakpointModel::aboutToReload() stops document line
tracking by all of the document's breakpoints. The stopping removes all
moving cursors and marks bound to the reloaded document. This prevents
problematic markChanged() signal emissions during the reload.

The aboutToReload() slot also connects a new slot
BreakpointModel::reloaded() to the reloaded() signal, unless there are
no breakpoints at the reloaded document's URL, in which case there is
nothing to be done. The reloaded() slot reinitializes all moving
cursors and breakpoint marks in the reloaded document at the
breakpoints' saved line numbers.

An unsolved issue is complete support for all possible reload
scenarios. The reloading process of a modified document is interrupted
by a question to the user. If the user selects an option other than
"Discard" (i.e. "Save" or "Cancel"), the breakpoints should not be
moved to their saved line numbers, but ought to remain in place
instead. Firstly, this requires determining the user choice in KDevelop
code, which is not trivial, because the question is asked by
ReadWritePart::queryClose() indirectly invoked from
KTextEditor::DocumentPrivate::documentReload() with several
intermediary functions. And secondly, adapting moving cursor and mark
changes to the user choice. These difficulties will be resolved in a
future commit.

Add a test function TestBreakpointModel::testDocumentReload(), which
verifies that reloading an unmodified document does not affect
breakpoints.