Specialize QTest::toString() for two Breakpoint enums
test_gdb and test_lldb QCOMPARE Breakpoint::BreakpointState enumerators
in many places. This enum is not and cannot be declared as Q_ENUM or
Q_ENUM_NS, because it is nested in the Breakpoint class, which does not
inherit QObject. When these QCOMPARE() checks fail, they do not print
actual and expected values, but only the uninformative general message
"Compared values are not the same".

test_breakpointmodel is going to QCOMPARE Breakpoint::BreakpointState
and Breakpoint::BreakpointKind enumerators in subsequent commits.
Implement QTest::toString() function template specializations for these
two similar enums in order to print at least numerical values of the
enumerators. Place the new specializations in
kdevplatform/tests/testhelpers.h, because this header already contains
many QTest::toString() specializations for custom KDevelop types and
because test_gdb, test_lldb and test_breakpointmodel already depend on
KDevPlatformTests.