Specialize QTest::toString() for IDocument::DocumentState
Two QCOMPARE(doc->state(), IDocument::Clean) calls in
TestBreakpointModel compare IDocument::DocumentState enumerators. This
enum is not and cannot be declared as Q_ENUM or Q_ENUM_NS, because it
is nested in the IDocument class, which does not inherit QObject. When
these QCOMPARE() checks fail, they do not print actual and expected
values, but only the uninformative general message "Compared values are
not the same". Implement a QTest::toString() function template
specialization for IDocument::DocumentState in order to print at least
numerical values of the enumerators.