Only reparse project if meson-info contents change
Currently there is a file watcher that reloads the project when the
meson-info.json file changes. If the project fails to configure,
kdevelop keeps attempting to reconfigure in a loop because the
meson-info.json file keeps getting rewritten. This change makes it
only reconfigure if the contents of the watched file are actually
changed avoiding the loop.

BUG: 482983
FIXED-IN: 5.14