BreakpointModel: prevent breakpoints in untitled documents
For an untitled/unsaved document, IDocument::url() returns
"file:///Untitled"* but KTextEditor::Document::url() returns an empty
URL. This difference causes bugs in BreakpointModel, because it assumes
the two functions are the same and uses whichever is more convenient.

Breakpoints in an untitled document are useless, because a debugger
cannot work with a KTextEditor buffer, only with a file on disk.

Implement a simple fix: prevent adding breakpoints to untitled
documents. First of all, do not enable breakpoint marks for an untitled
document in textDocumentCreated(). This leaves only bookmarks in the
untitled document border mark user interface, and therefore ensures
that markChanged() is never invoked with a breakpoint mark in an
untitled document.

In order to allow adding breakpoints as soon as an untitled document is
saved, enable breakpoint marks in documentSaved(). Optimize
documentSaved() along the way by returning early if a non-text document
is saved.

Prevent toggleBreakpoint() and markContextMenuRequested() from adding a
breakpoint to an untitled document by early-returning if breakpoint
marks are not enabled.