git plugin: fix parsing of commits that are tagged or tip of a branch
The tip of my branch was missing, and the contents of other commits (e.g. logs)
were showing up merged. It turns out that those commits would show up in the log
like this:

  commit <SHA1> (tag: barfoo, HEAD -> foobar, origin/foobar)

I set a different format for "git log", which also included the PGP signature
state which would then show up in the "author" field of the log, usually as
"[undefined N]" when commits were not signed.

Tell git to output the commits in the expected "medium" format and omit those
decorations.