BreakpointWidget: QTreeView => QTableView
Mostly revert 59420a7b341b6901be7a71a5b5a09cfb2d5cb513, but also hide
the table grid. This way the appearance is almost the same.

One advantage of the table view is that double-clicking on the last
placeholder row or pressing the Plus key while that row is selected no
longer tries to expand the item, which causes an assertion failure in
PlaceholderItemProxyModel::rowCount(). The assertion failure can also be
triggered by pressing the Asterisk key while any row of the focused
Breakpoints table is selected. An alternative fix is to call
setItemsExpandable(false) on the tree view, but if the table view looks
just as well or better, it is more direct, simple and efficient, and
carries less risk of future bugs like that.