BreakpointModel: skip setting up moving cursors in untitled documents
There is no need to pass an untitled document to
setupDocumentBreakpoints(), because we forbid breakpoints in such
documents. A breakpoint's empty URL value means "none", so
setupDocumentBreakpoints() could create wrong moving cursors for
expression breakpoints in an untitled document.

An untitled document cannot be reloaded, so reloaded() never passes
such a document to setupDocumentBreakpoints().