MIParser: use const to prevent detaching
Also eliminate the intermediate QByteArray representation that is used
once, pass the QByteArray instead of its raw data to QString::fromUtf8()
and make a local int variable const as well.

This code was practically not changed since the time when the UTF-8
conversion was introduced in 5cbe55083839d0baf9e92e8326cc24b31332e365.