gdb: pretty-print QByteArray as string, not const char*
Returning a const char* from QByteArrayPrinter.to_string() shows the HEX
pointer value followed by the string in double quotes in Variables tool
view and debugger tooltips. The HEX pointer value inside a QByteArray is
an irrelevant implementation detail. Convert the const char* to string
in order to show only the string in double quotes in the UI.

The first optional argument to gdb.Value.string() is `encoding`.
`encoding = 'UTF-8'` could be passed, but QByteArray does not mandate or
recommend any encoding. Also on my system passing this argument does not
seem to make any difference: some non-ASCII characters are and others
are not correctly displayed in the UI, in any case. So omit the encoding
argument and let GDB try and figure out the encoding itself.