BreakpointModel: work around another empty URL assertion failure
The assertion failure can be reproduced by combining steps listed in the
added comments.

If the document's URL is not empty, the assertion in markChanged() does
not fail. However, when KTextEditor::DocumentPrivate::clearMarks()
removes a leftover out-of-sync breakpoint mark, the resulting invocation
of BreakpointModel::markChanged() can accidentally find some unrelated
inactive (with document line tracking disabled for some reason)
breakpoint on the out-of-sync breakpoint mark's line and remove it. But
such a coincidence is highly unlikely to occur in practice. And such an
accidental removal of an inactive breakpoint should be prevented by
doing nothing if markChanged(document, mark, MarkRemoved) finds no
breakpoint with a valid moving cursor on document's mark.line.