Breakpoint: clear unsupported URL read from config
Otherwise, if an unsupported URL ends up in KDevelop's config file
sessionrc and the user activates an unsupported-URL breakpoint's row in
Breakpoints tool view, one of the assertions at the beginning of
DocumentControllerPrivate::openDocumentInternal() can fail.

Assign -1 to Breakpoint::m_line instead of reading the line from config
if the breakpoint's URL is empty (possibly unsupported and cleared).
This prevents constructing a breakpoint location string using an empty
URL in Breakpoint::data(), which uses a code breakpoint's expression
instead of its URL and line if and only if the line equals -1.