test_lldb: fix testStack*()
With LLDB versions 16.0.6 and 17.0.6 on GNU/Linux, a call stack contains
a third last frame named "___lldb_unnamed_symbol3264" between "main" and
"__libc_start_main". With LLDB version 17.0.6 on FreeBSD, the third last
frame is "main" and the penultimate frame's name is "__libc_start1".

Do not check system-specific frame names and verify the minimum expected
stack frame counts, which are equal to the corresponding counts in
test_gdb, because GDB/MI does not report system stack frames. This
should make the tests more robust in the face of LLDB changes and less
platform-dependent.