debugger: don't use invalid actual location of a lldb-mi breakpoint
Unlike GDB/MI, lldb-mi sends the "fullname" and "line" fields for a
pending breakpoint. But these sent field values are invalid: "??/??" and
(one-based) "0" respectively.

If the "fullname" and "line" fields are present,
MIBreakpointController::updateFromDebugger() converts "fullname" to a
local-file URL and passes it to Breakpoint::setLocation(). The invalid
"fullname" value is thus converted to an unsupported local-file URL with
a relative path "file://??/??", which makes the assertion
isSupportedBreakpointUrl(url) in Breakpoint::setLocation() fail since
the grandparent commit.

Before the grandparent commit,
MIBreakpointController::updateFromDebugger() also wrongly replaced the
original breakpoint location with the invalid URL and line values, but
that did not immediately trigger an assertion failure. Instead, the
original locations of all pending breakpoints were lost as soon as a
LLDB debug session started.

Treat the invalid "line" field value of 0 in the same way as missing
"line" field to consistently keep the original location of a pending
breakpoint intact with both GDB/MI and lldb-mi.

MIBreakpointController::createFromDebugger() also uses the "fullname"
and "line" values, but does not pass them to Breakpoint::setLocation(),
and its existing "original-location" fallback works correctly then.
Therefore, rejecting the invalid actual location in createFromDebugger()
is merely an optimization.

The added test function tests breakpoints on lines that consist of no-op
code. GDB and LLDB move some of these breakpoints to the closest next
stoppable line. Others remain on their no-op lines and are permanently
pending. Therefore, LldbTest::testBreakpointsOnNoOpLines() triggers the
assertion failure in the absence of this commit's fix and fails before
the grandparent commit (because a pending breakpoint's line number is
replaced with (one-based) 0).

Expose introduced helper functions in testhelper.h. They are currently
used only in testhelper.cpp, but can be reused in test_gdb and test_lldb
in the future.