Revert "work around bug in kLineEdit"
The worked around Bug 388798 was fixed in KCompletion version 5.43.
KDevelop requires KDE Frameworks version 5.91 or later. I have verified
the absence of crash when switching between Auto-detected and Manual
compilers on the Language Support=>C/C++ Compilers tab of the
Configure - KDevelop dialog (at this commit, with the workaround
reverted).

This reverts commit e6ee67c23ae333d2ea4129bf507c5a389684c960.