Actually show status messages in the status bar
IdealButtonBarWidget adds stretchable space to its layout, which grows
to occupy all free horizontal space. This prevents ever displaying
status messages that are not tied to progress items and are passed to
QStatusBar::showMessage(). Remove the stretchable space to show the
messages.

Fix the progress widget's width. Otherwise, the progress bar grows to an
ugly great length at the expense of status messages that can actually
make use of the horizontal space.