Fix warning about Plugin ID Metadata
This fixes warnings of the type:

kf.coreaddons: The plugin
"/home/mvo/kde/usr/lib/x86_64-linux-gnu/plugins/kdevplatform/60/kdevbazaar.so"
explicitly states an 'Id' in the embedded metadata. This value should be
removed, the resulting pluginId will not be affected by it

By removing the Id from the .json files in the relevant plugins. The
qmljs plugin is not ported to KF6 and is disabled, so its ID is not
removed.

Also remove the Id from the .json file generated by the kdevelop plugin
template.