Don't show background parser's messages on the status bar
Background parser's messages are clearly aimed at the status of its
progress item. But when the progress item is absent, these status
messages are shown on the status bar for an unlimited time. Prevent this
annoyance by introducing a special negative timeout value for
IStatus::showMessage() and handling it accordingly in StatusBar.

A status message with a negative timeout is never displayed on the
status bar, but is still inserted into StatusBar::m_messages in order to
allow StatusBar::showProgress() to transfer it to a new progress item's
status.