StatusBar: don't subtract zero from message timeouts
m_time.elapsed() equals 0 in StatusBar::updateMessage() if a new status
message arrives very soon after the previous one. This happens in
practice when the "Process exited" status message is immediately
followed by the "Exited normally" status message at the end of a debug
session. An integer equality check is much faster than iterating over a
QHash, so this is an optimization.