StatusBar: transfer existing status message to new progress item
If a corresponding progress item exists, StatusBar::showMessage()
assigns its message argument to the progress item's status in preference
to displaying it as part of the aggregate status message. If the
corresponding progress item is created after a status message is passed
to StatusBar::showMessage(), the item's initial status is empty and the
message remains separate. Transfer an existing status message (if any)
to the corresponding new progress item, where it is better categorized
and more useful.