Show an alert when a breakpoint or a watchpoint is hit
The default action for BreakpointHit event is Taskbar. This action got
removed in KF6:
https://commits.kde.org/knotifications/d91844f90ca4606399c2b6cee73cd7cd91d9d2f3
Breakpoints are hit often, so just remove the no longer existent
BreakpointHit default action rather than set it to Popup or Sound. Do
not remove the BreakpointHit event itself, because users may want to
assign some action to it.

QApplication::alert() is equivalent to the KF5 Taskbar action. So call
this function when a breakpoint is hit to preserve the default KF5
behavior. Show an alert when a watchpoint is hit as well - does not hurt
and could be particularly useful if the user disables the default
WatchpointHit action Popup.

This change is inspired by an old merge request and comments under it:
https://invent.kde.org/kdevelop/kdevelop/-/merge_requests/86